+++ src/main/java/com/trackery/trackeryfrontserver/domain/mypage/controller/MyPageController.java
@@ -3,6 +3,7 @@ package com.trackery.trackeryfrontserver.domain.mypage.controller;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 /**
  * packageName    : com.trackery.trackeryfrontserver.domain.mypage.controller
@@ -18,9 +19,16 @@ import org.springframework.web.bind.annotation.GetMapping;
  */
 @Slf4j
 @Controller
+@RequestMapping("/mypage")
 public class MyPageController {
-    @GetMapping("/mypage")
+    @GetMapping("")
     public String myPage() {
         return "mypage/mypage-content";
     }
+
+    @GetMapping("/update-user-info-modal")
+    public String  updateUserInfoModal() {
+        return "mypage/update-user-info-modal";
+    }
+
 }
