+++ src/main/resources/static/mypage/css/update-user-info.css
@@ -1,0 +1,6 @@
+.update-user-info-modal-container{
+    background-color: #0a66c2;
+    display: flex;
+    justify-content: right;
+    width: 30%;
+}
