+++ src/main/resources/static/mypage/js/MyPage.js
@@ -1,3 +1,13 @@
+import {loadModal} from "/module/modal/ModalUtil.js";
+
+document.addEventListener("DOMContentLoaded", async function() {
+    try {
+        await loadModal("/mypage/update-user-info-modal", document.getElementsByClassName("update-user-info-modal-container")[0], "/mypage/js/UpdateUserInfo.js");
+    } catch (error) {
+        console.error("모달 로딩 중 오류 발생:", error)
+    }
+})
+
 document.addEventListener("DOMContentLoaded", function () {
     fetch("/api/users/details", {
         method: "GET",
