+++ src/main/resources/templates/mypage/mypage-content.html
@@ -24,6 +24,10 @@
                 </div>
             </div>
 
+            <div class="update-user-info-modal-container">
+
+            </div>
+
             <!--공개 앨범-->
             <div class="published-album-container">
                 <p id="published-album-title">공개 앨범</p>
@@ -31,7 +35,7 @@
         </div>
     </section>
     <div layout:fragment="scripts">
-        <script src="/mypage/js/mypage.js"></script>
+        <script type="module" src="/mypage/js/mypage.js"></script>
     </div>
 </body>
 </html>
